requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gTTS==2.5.1
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
playsound==1.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.8
pyttsx3==2.90
pywin32==306
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
